requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
bidict==0.23.1
blinker==1.9.0
bson==0.5.10
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
deprecation==2.1.0
dnspython==2.8.0
eventlet==0.33.3
fastapi==0.123.3
Flask==2.3.2
Flask-SocketIO==5.3.4
fsspec==2026.2.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
motor==3.7.1
multidict==6.7.1
packaging==26.0
postgrest==2.28.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.8.0
pymongo==4.15.4
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.8.0
realtime==2.28.0
requests==2.32.5
rich==14.3.3
simple-websocket==1.1.0
six==1.17.0
starlette==0.50.0
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.38.0
websockets==15.0.1
Werkzeug==2.3.6
wsproto==1.3.2
yarl==1.23.0
zstandard==0.25.0
